Use context manager for cairo save/restore
